feat(core): Add PBKDF2 support
Extend NSData+Crypto to support PBKDF2 with SHA1 HMAC as dovecot
is using it since v2.3.0.

The format hashed passwords is {PBKDF2}$1$<salt>$<rounds>$<hashed value in hex>

Depending on the SSL library, either GnuTLS or OpenSSL functions
are used for the heavy-lifting of the actual PKCS#5 PBKDF2
implementation. Note that to work with GnuTLS, version >= 3.6.13 is
required.